refactor: move locale switcher to components
